+++ src/main/kotlin/io/exam/intellij/plugin/action/DebugExamplesAction.kt
@@ -1,7 +1,8 @@
 package io.exam.intellij.plugin.action
 
+import com.intellij.execution.Executor
 import com.intellij.execution.executors.DefaultDebugExecutor
 
 class DebugExamplesAction : RunExamplesAction() {
-    override fun getExecutor() = DefaultDebugExecutor.getDebugExecutorInstance().id
+    override fun getExecutor(): Executor = DefaultDebugExecutor.getDebugExecutorInstance()
 }